src/requirements.txt
flet
flet-web
spotipy
sqlalchemy
python-dotenv
psycopg2-binary
aiosqlite
asyncpg
greenlet
